+++ samples/Oragon.AdaptivePool.RabbitMQ.Sample.LiveDashboard/Oragon.AdaptivePool.RabbitMQ.Sample.LiveDashboard.Web/Program.cs
@@ -1,6 +1,7 @@
 using Oragon.AdaptivePool.RabbitMQ.Sample.LiveDashboard.Web.Components;
 using Oragon.AdaptivePool.RabbitMQ.Sample.LiveDashboard.Web.Live;
 using Oragon.AdaptivePool.RabbitMQ.DependencyInjection;
+using RabbitMQ.Client;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -17,7 +18,7 @@ builder.Services.AddAdaptiveConnectionPool(
     name: LivePoolNames.PoolName,
     configureFactory: factory => factory.Uri = new Uri(rabbitMqConnectionString),
     configurePool: pool => pool
-        .WithBounds(min: 1, max: 200, initial: 1)
+        .WithBounds(min: 1, max: 10, initial: 1)
         .WithIdleTimeout(TimeSpan.FromSeconds(3))
         .WithSweepInterval(TimeSpan.FromSeconds(1))
         .WithShrinkOnUtilizationPercent(0.60)
@@ -29,13 +30,19 @@ builder.Services.AddAdaptiveChannelPool(
     name: LivePoolNames.PoolName,
     connectionPoolName: LivePoolNames.PoolName,
     configurePool: pool => pool
-        .WithBounds(min: 0, max: 200, initial: 0)
+        .WithBounds(min: 0, max: 50, initial: 0)
         .WithIdleTimeout(TimeSpan.FromSeconds(3))
         .WithSweepInterval(TimeSpan.FromSeconds(1))
         .WithShrinkOnUtilizationPercent(0.60)
         .WithShrinkTargetUtilizationPercent(0.75)
         .WithShrinkBatchSize(64)
         .WithShrinkCooldownWindows(1)
+        // Throughput test mode: publisher confirms/tracking disabled.
+        .WithChannelOptions(new CreateChannelOptions(
+            publisherConfirmationsEnabled: false,
+            publisherConfirmationTrackingEnabled: false,
+            outstandingPublisherConfirmationsRateLimiter: null,
+            consumerDispatchConcurrency: 1))
         .WithMaxChannelsPerConnection(16));
 
 builder.Services.AddSingleton<LiveLoadController>();